% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw}
\alias{rw}
\title{A markov sampler using Rcpp}
\usage{
rw(sigma, x0, N)
}
\arguments{
\item{sigma}{the variance}

\item{x0}{the Initial value}

\item{N}{Number of random numbers}
}
\value{
Returns random sequence and rejection probability
}
\description{
A markov sampler using Rcpp
}
\examples{
\dontrun{
rw(5,7,500)
}
}
