\name{SC19072-package}
\alias{SC19072-package}
\docType{package}
\title{
  Multi-attribute Decision Based on Information Entropy
}
\description{
  R  versions of select and savage.select function and Rcpp version of rw are provided in this package
}
\author{
Du Xianyang

Maintainer: Du Xianyang <ustcxydu@mail.ustc.edu.cn>
}
\keyword{ package }
